Drop the m parameter from HasIn.
Without this, ghc was unable to compute any relationship between Has eff m and Has eff (t m) when Signature (t m) = Signature m.

This makes HasIn more or less Member with its arguments in the opposite order, but at least the error messages suggest Control.Algebra this way, and the name relates to the constraint the user wrote.